refactor: extract large module methods to reduce file sizes
Split rns_menu_mixin.py (1,904 → 1,041 lines):
- Extract config management to rns_config_mixin.py (425 lines)
- Extract diagnostics/tool execution to rns_diagnostics_mixin.py (505 lines)
- RNSMenuMixin now inherits from both new mixins

Split rns_bridge.py (1,652 → 1,525 lines):
- Extract MeshCore bridge processing to meshcore_bridge_mixin.py (169 lines)
- RNSMeshtasticBridge now inherits from MeshCoreBridgeMixin

Update CLAUDE.md file size audit:
- Add service_menu_mixin.py (1,611 lines) to tracking
- Update all line counts to current values
- Add refactoring history entries

All 1,523 tests pass.
