+++ src/launcher_tui/rns_config_mixin.py
@@ -1,0 +1,425 @@
+"""
+RNS Config Mixin - Reticulum configuration management.
+
+Extracted from rns_menu_mixin.py to reduce file size per CLAUDE.md guidelines.
+Handles config viewing, editing, validation, deployment, and migration.
+"""
+
+import logging
+import os
+import re
+import shutil
+import subprocess
+from pathlib import Path
+from typing import Optional
+
+logger = logging.getLogger(__name__)
+
+from utils.paths import get_real_user_home, ReticulumPaths
+from backend import clear_screen
+
+
+class RNSConfigMixin:
+    """Mixin providing RNS configuration management functionality."""
+
+    @staticmethod
+    def _is_root_owned_rns_config(config_path: Path) -> bool:
+        """Check if the RNS config is in a root-only location (/root/)."""
+        try:
+            return str(config_path.resolve()).startswith('/root/')
+        except OSError:
+            return str(config_path).startswith('/root/')
+
+    def _migrate_rns_config_to_etc(self, source: Path) -> bool:
+        """Migrate RNS config from root-owned location to /etc/reticulum/config.
+
+        Copies the config to /etc/reticulum/config (system-wide, preferred location),
+        sets world-readable permissions, and renames the old file to avoid confusion.
+
+        Returns True if migration succeeded.
+        """
+        target = Path('/etc/reticulum/config')
+        if target.exists():
+            self.dialog.msgbox(
+                "Cannot Migrate",
+                f"Config already exists at:\n  {target}\n\n"
+                f"Remove it first if you want to migrate from:\n  {source}"
+            )
+            return False
+        try:
+            target.parent.mkdir(parents=True, exist_ok=True)
+            shutil.copy2(str(source), str(target))
+            target.chmod(0o644)
+            # Rename old config so rnsd picks up the /etc/ one
+            backup = source.with_suffix('.migrated')
+            source.rename(backup)
+            return True
+        except (OSError, PermissionError) as e:
+            self.dialog.msgbox("Error", f"Failed to migrate config:\n{e}")
+            return False
+
+    def _deploy_rns_template(self) -> Optional[Path]:
+        """Deploy RNS template to /etc/reticulum/config (system-wide).
+
+        Returns the path where the config was deployed, or None on failure.
+        """
+        template = Path(__file__).parent.parent.parent / 'templates' / 'reticulum.conf'
+        if not template.exists():
+            return None
+
+        # Always deploy to /etc/reticulum/ (system-wide, first in search order)
+        target = Path('/etc/reticulum/config')
+        if target.exists():
+            self.dialog.msgbox(
+                "Config Exists",
+                f"Config already exists at:\n  {target}\n\n"
+                f"Use 'Edit Reticulum Config' to modify it."
+            )
+            return None
+        try:
+            target.parent.mkdir(parents=True, exist_ok=True)
+            shutil.copy2(str(template), str(target))
+            target.chmod(0o644)  # World-readable so all users and rnsd can read it
+            return target
+        except (OSError, PermissionError) as e:
+            self.dialog.msgbox("Error", f"Failed to deploy config:\n{e}")
+            return None
+
+    def _view_rns_config(self):
+        """View current Reticulum config."""
+        clear_screen()
+        print("=== Reticulum Configuration ===\n")
+
+        config_path = ReticulumPaths.get_config_file()
+
+        if config_path.exists():
+            # Warn if config is in root-only location
+            if self._is_root_owned_rns_config(config_path):
+                print(f"Config: {config_path}")
+                print(f"  ** This config is in /root/ - not editable without sudo **")
+                print(f"  ** Use 'Edit Reticulum Config' to migrate to /etc/reticulum/ **\n")
+            else:
+                print(f"Config: {config_path}\n")
+            try:
+                content = config_path.read_text()
+                print(content)
+
+                # Show validation warnings inline
+                issues = self._validate_rns_config_content(content)
+                if issues:
+                    print("\n--- Config Issues ---")
+                    for issue in issues:
+                        print(f"  ! {issue}")
+            except PermissionError:
+                print(f"Permission denied reading {config_path}")
+                print(f"Try: sudo cat {config_path}")
+        else:
+            print(f"No Reticulum config found at: {config_path}")
+            user_home = get_real_user_home()
+            print(f"\nMeshForge checks (in order):")
+            print(f"  1. /etc/reticulum/config  (system-wide, preferred)")
+            print(f"  2. {user_home}/.config/reticulum/config")
+            print(f"  3. {user_home}/.reticulum/config")
+            if os.geteuid() == 0 and os.environ.get('SUDO_USER'):
+                print(f"\nNote: rnsd (running as root) uses /root/.reticulum/config")
+                print(f"  For shared use, deploy to /etc/reticulum/config")
+            print(f"\nTo create: use 'Edit Reticulum Config' to deploy template")
+            print(f"Template:  templates/reticulum.conf")
+
+        # Show Meshtastic_Interface plugin status
+        plugin_path = ReticulumPaths.get_interfaces_dir() / 'Meshtastic_Interface.py'
+        print(f"\n--- Meshtastic Interface Plugin ---")
+        if plugin_path.exists():
+            print(f"  Installed: {plugin_path}")
+            print(f"  Size: {plugin_path.stat().st_size} bytes")
+        else:
+            print(f"  NOT INSTALLED")
+            print(f"  Expected at: {plugin_path}")
+            print(f"  Source: https://github.com/landandair/RNS_Over_Meshtastic")
+            print(f"  Use 'Install Meshtastic Interface' from the RNS menu to install.")
+
+        self._wait_for_enter()
+
+    def _edit_rns_config(self):
+        """Edit Reticulum config with available editor. Deploys template if no config exists."""
+        config_path = ReticulumPaths.get_config_file()
+
+        if not config_path.exists():
+            # Offer to deploy from template to /etc/reticulum/config (system-wide)
+            template = Path(__file__).parent.parent.parent / 'templates' / 'reticulum.conf'
+
+            if template.exists():
+                target = Path('/etc/reticulum/config')
+                if self.dialog.yesno(
+                    "Deploy Reticulum Config",
+                    f"No Reticulum config found.\n\n"
+                    f"Deploy template to:\n  {target}\n\n"
+                    f"This sets up RNS with:\n"
+                    f"  - share_instance = Yes (required for rnstatus)\n"
+                    f"  - AutoInterface (local network discovery)\n"
+                    f"  - Meshtastic_Interface on port 4403\n\n"
+                    f"You can edit it after deployment."
+                ):
+                    deployed = self._deploy_rns_template()
+                    if deployed:
+                        config_path = deployed
+                    else:
+                        return
+                else:  # User said No
+                    return
+            else:
+                self.dialog.msgbox(
+                    "No Config",
+                    "No Reticulum config found and template missing.\n\n"
+                    "Install RNS first: pipx install rns\n"
+                    "Then run rnsd once to generate default config."
+                )
+                return
+
+        # If config is in /root/, offer to migrate to /etc/reticulum/
+        if self._is_root_owned_rns_config(config_path):
+            if self.dialog.yesno(
+                "Migrate Config",
+                f"Config is at:\n  {config_path}\n\n"
+                f"This location requires root access to edit.\n\n"
+                f"Migrate to /etc/reticulum/config?\n"
+                f"(System-wide location, accessible by rnsd and all users)"
+            ):
+                if self._migrate_rns_config_to_etc(config_path):
+                    config_path = Path('/etc/reticulum/config')
+                    self.dialog.msgbox(
+                        "Migrated",
+                        f"Config moved to: {config_path}\n\n"
+                        f"Restart rnsd to apply:\n"
+                        f"  sudo systemctl restart rnsd"
+                    )
+                # If migration failed, continue with original path
+
+        # Find editor
+        editor = None
+        for cmd in ['nano', 'vim', 'vi']:
+            if shutil.which(cmd):
+                editor = cmd
+                break
+
+        if not editor:
+            self.dialog.msgbox("Error", "No text editor found (nano, vim, vi)")
+            return
+
+        subprocess.run([editor, str(config_path)], timeout=None)
+
+        # After editing, check for config divergence between user and root
+        self._check_rns_config_divergence(config_path)
+
+    def _check_rns_config_divergence(self, edited_path: Path):
+        """Check if edited config differs from root/system config that rnsd actually uses.
+
+        When running with sudo, user edits /home/user/.reticulum/config but
+        rnsd (running as root) reads /root/.reticulum/config. The configs
+        silently diverge, causing the user's changes to have no effect.
+
+        This check warns the user and offers to sync.
+        """
+        # Only relevant when running as root/sudo
+        if os.geteuid() != 0:
+            return
+
+        # Find where rnsd actually reads its config
+        # rnsd systemd service runs as root, so it reads from one of:
+        root_configs = [
+            Path('/etc/reticulum/config'),
+            Path('/root/.config/reticulum/config'),
+            Path('/root/.reticulum/config'),
+        ]
+
+        # Skip if edited path is already a root/system path
+        edited_str = str(edited_path)
+        if edited_str.startswith('/root/') or edited_str.startswith('/etc/'):
+            return
+
+        for root_config in root_configs:
+            if root_config.exists() and root_config != edited_path:
+                # Compare contents
+                try:
+                    user_content = edited_path.read_text()
+                    root_content = root_config.read_text()
+
+                    if user_content != root_content:
+                        if self.dialog.yesno(
+                            "Config Divergence Detected",
+                            f"WARNING: Your edited config:\n"
+                            f"  {edited_path}\n\n"
+                            f"differs from the config rnsd uses:\n"
+                            f"  {root_config}\n\n"
+                            f"rnsd runs as root and reads {root_config}.\n"
+                            f"Your changes won't take effect until synced.\n\n"
+                            f"Copy your config to {root_config}?"
+                        ):
+                            try:
+                                # Backup root config first
+                                backup = root_config.with_suffix('.config.bak')
+                                if root_config.exists():
+                                    shutil.copy2(str(root_config), str(backup))
+                                shutil.copy2(str(edited_path), str(root_config))
+                                self.dialog.msgbox(
+                                    "Config Synced",
+                                    f"Copied to: {root_config}\n"
+                                    f"Backup at: {backup}\n\n"
+                                    f"Restart rnsd to apply:\n"
+                                    f"  sudo systemctl restart rnsd"
+                                )
+                            except Exception as e:
+                                self.dialog.msgbox(
+                                    "Sync Failed",
+                                    f"Could not copy config: {e}\n\n"
+                                    f"Manual fix:\n"
+                                    f"  sudo cp {edited_path} {root_config}\n"
+                                    f"  sudo systemctl restart rnsd"
+                                )
+                except (OSError, subprocess.SubprocessError) as e:
+                    logger.debug("RNS config apply failed: %s", e)
+                return  # Only check the first existing root config
+
+    def _validate_rns_config_content(self, content: str) -> list:
+        """Validate RNS config content and return list of issues found.
+
+        Checks for common misconfigurations that cause rnstatus/rnpath failures:
+        - Missing [reticulum] section
+        - Missing share_instance = Yes (required for client apps to connect)
+        - No interfaces configured
+        - No Meshtastic_Interface (needed for mesh bridging)
+        - Meshtastic_Interface.py plugin not installed
+        """
+        issues = []
+        content_lower = content.lower()
+
+        # Check [reticulum] section exists
+        if '[reticulum]' not in content_lower:
+            issues.append("Missing [reticulum] section")
+
+        # Check share_instance (required for rnstatus/rnpath to connect to rnsd)
+        has_share = False
+        for line in content.split('\n'):
+            stripped = line.strip().lower()
+            if stripped.startswith('#'):
+                continue
+            if 'share_instance' in stripped:
+                if 'yes' in stripped or 'true' in stripped:
+                    has_share = True
+                break
+        if not has_share:
+            issues.append("share_instance not set to Yes (rnstatus/client apps won't connect)")
+
+        # Check for at least one active interface
+        has_interface = False
+        has_meshtastic = False
+        for line in content.split('\n'):
+            stripped = line.strip()
+            if stripped.startswith('#'):
+                continue
+            if stripped.startswith('[[') and stripped.endswith(']]'):
+                has_interface = True
+            if 'meshtastic_interface' in stripped.lower() and 'type' in stripped.lower():
+                has_meshtastic = True
+
+        if not has_interface:
+            issues.append("No interfaces configured")
+
+        # Check Meshtastic_Interface status: config reference + plugin file
+        plugin_path = ReticulumPaths.get_interfaces_dir() / 'Meshtastic_Interface.py'
+        plugin_installed = plugin_path.exists()
+
+        if not has_meshtastic and not plugin_installed:
+            issues.append("No Meshtastic_Interface configured (needed for mesh bridging)")
+        elif has_meshtastic and not plugin_installed:
+            issues.append(
+                f"Meshtastic_Interface.py plugin not installed at "
+                f"{ReticulumPaths.get_interfaces_dir()}/\n"
+                f"    Install from: https://github.com/landandair/RNS_Over_Meshtastic"
+            )
+
+        return issues
+
+    def _check_rns_setup(self) -> bool:
+        """Check RNS setup and offer to fix common issues.
+
+        Available via 'Check RNS Setup' menu item. Returns True if setup
+        looks OK or user chose to continue anyway, False if user wants
+        to go back.
+        """
+        config_path = ReticulumPaths.get_config_file()
+
+        if not config_path.exists():
+            template = Path(__file__).parent.parent.parent / 'templates' / 'reticulum.conf'
+            if template.exists():
+                target = Path('/etc/reticulum/config')
+                if self.dialog.yesno(
+                    "RNS Not Configured",
+                    f"No Reticulum config found.\n\n"
+                    f"RNS tools (rnstatus, rnpath) and the gateway bridge\n"
+                    f"require a config file to function.\n\n"
+                    f"Deploy MeshForge template to:\n"
+                    f"  {target}\n\n"
+                    f"(Sets up shared instance + Meshtastic bridge)"
+                ):
+                    deployed = self._deploy_rns_template()
+                    if deployed:
+                        self.dialog.msgbox(
+                            "Config Deployed",
+                            f"Deployed to: {deployed}\n\n"
+                            f"Restart rnsd to apply:\n"
+                            f"  sudo systemctl restart rnsd"
+                        )
+                        config_path = deployed
+            return True  # Continue to menu either way
+
+        # Config exists - check if it's in a root-only location
+        if self._is_root_owned_rns_config(config_path):
+            if self.dialog.yesno(
+                "Config in /root/",
+                f"RNS config found at:\n  {config_path}\n\n"
+                f"This location requires root access to edit.\n\n"
+                f"Migrate to /etc/reticulum/config?\n"
+                f"(System-wide location, accessible by all users)"
+            ):
+                if self._migrate_rns_config_to_etc(config_path):
+                    config_path = Path('/etc/reticulum/config')
+                    self.dialog.msgbox(
+                        "Migrated",
+                        f"Config moved to: {config_path}\n\n"
+                        f"Restart rnsd to apply:\n"
+                        f"  sudo systemctl restart rnsd"
+                    )
+
+        # Config exists - validate it
+        try:
+            content = config_path.read_text()
+            issues = self._validate_rns_config_content(content)
+            if issues:
+                msg = f"Config: {config_path}\n\nIssues found:\n"
+                for issue in issues:
+                    msg += f"  - {issue}\n"
+                msg += f"\nUse 'Edit Reticulum Config' to fix these issues."
+                self.dialog.msgbox("RNS Config Issues", msg)
+        except PermissionError:
+            self.dialog.msgbox(
+                "Permission Denied",
+                f"Cannot read config at:\n  {config_path}\n\n"
+                f"Run MeshForge with sudo to access this file,\n"
+                f"or use 'Edit Reticulum Config' to migrate it."
+            )
+
+        # Check for Meshtastic_Interface.py plugin (separate from config validation)
+        if not self._check_meshtastic_plugin():
+            if self.dialog.yesno(
+                "Meshtastic Interface Plugin Missing",
+                "The Meshtastic_Interface.py plugin is not installed.\n\n"
+                "This plugin is required for bridging RNS over\n"
+                "Meshtastic LoRa mesh networks.\n\n"
+                f"Expected at:\n"
+                f"  {ReticulumPaths.get_interfaces_dir()}/Meshtastic_Interface.py\n\n"
+                "Download and install it now?"
+            ):
+                self._install_meshtastic_interface_plugin()
+
+        return True
