+++ src/launcher_tui/rns_menu_mixin.py
@@ -3,6 +3,8 @@ RNS Menu Mixin - Reticulum Network Stack menu handlers.
 
 Extracted from main.py to reduce file size per CLAUDE.md guidelines.
 Sniffer methods further extracted to rns_sniffer_mixin.py.
+Config methods extracted to rns_config_mixin.py.
+Diagnostics methods extracted to rns_diagnostics_mixin.py.
 """
 
 import logging
@@ -16,6 +18,8 @@ from typing import Optional
 logger = logging.getLogger(__name__)
 
 from rns_sniffer_mixin import RNSSnifferMixin
+from rns_config_mixin import RNSConfigMixin
+from rns_diagnostics_mixin import RNSDiagnosticsMixin
 
 # Import centralized path utility - SINGLE SOURCE OF TRUTH for all paths
 # See: utils/paths.py (ReticulumPaths, get_real_user_home)
@@ -42,10 +46,12 @@ detect_rnsd_config_drift, _HAS_CONFIG_DRIFT = safe_import(
 )
 
 
-class RNSMenuMixin(RNSSnifferMixin):
+class RNSMenuMixin(RNSSnifferMixin, RNSConfigMixin, RNSDiagnosticsMixin):
     """Mixin providing RNS/Reticulum menu functionality.
 
     Inherits sniffer methods from RNSSnifferMixin.
+    Inherits config methods from RNSConfigMixin.
+    Inherits diagnostics methods from RNSDiagnosticsMixin.
     """
 
     def _rns_menu(self):
@@ -539,259 +545,6 @@ class RNSMenuMixin(RNSSnifferMixin):
         self._repair_rns_shared_instance()
         self._wait_for_enter()
 
-    def _rns_diagnostics(self):
-        """Run comprehensive RNS diagnostics."""
-        clear_screen()
-        print("=== RNS Diagnostics ===\n")
-
-        if not _HAS_RNS_COMMANDS:
-            print("RNS commands module not available.")
-            print("Run from MeshForge root: sudo python3 src/launcher_tui/main.py")
-            self._wait_for_enter()
-            return
-
-        # 1. Service status
-        print("[1/5] Checking rnsd service...")
-        status = get_status()
-        status_data = status.data or {}
-        running = status_data.get('rnsd_running', False)
-        service_state = status_data.get('service_state', '')
-        print(f"  rnsd: {'RUNNING' if running else 'NOT RUNNING'}")
-        if status_data.get('rnsd_pid'):
-            print(f"  PID: {status_data['rnsd_pid']}")
-        if service_state:
-            print(f"  State: {service_state}")
-
-        # Detect NomadNet conflict (common cause of rnsd crash-loops)
-        nomadnet_conflict = self._check_nomadnet_conflict()
-        if nomadnet_conflict:
-            print(f"  NomadNet: RUNNING (port conflict!)")
-        if service_state == 'failed' or (not running and nomadnet_conflict):
-            print("")
-            if nomadnet_conflict:
-                print("  WARNING: NomadNet is holding the RNS shared instance port.")
-                print("  rnsd cannot bind port 37428 while NomadNet is running.")
-                print("  Fix: stop NomadNet first, or disable rnsd and let NomadNet")
-                print("  serve as the shared instance.")
-            elif service_state == 'failed':
-                print("  WARNING: rnsd has crashed. Check logs:")
-                print("    sudo journalctl -u rnsd -n 30")
-
-        # 2. Config check
-        print("\n[2/5] Checking configuration...")
-        config_exists = status_data.get('config_exists', False)
-        print(f"  Config: {'found' if config_exists else 'MISSING'}")
-        if config_exists:
-            iface_count = status_data.get('interface_count', 0)
-            print(f"  Interfaces: {iface_count}")
-
-        # 3. Identity check
-        print("\n[3/5] Checking identity...")
-        identity_exists = status_data.get('identity_exists', False)
-        print(f"  Gateway identity: {'found' if identity_exists else 'not created'}")
-        config_dir = ReticulumPaths.get_config_dir()
-        rns_identity = config_dir / 'identity'
-        print(f"  RNS identity: {'found' if rns_identity.exists() else 'not created'}")
-
-        # 4. Full connectivity check
-        print("\n[4/5] Running connectivity check...")
-        conn = check_connectivity()
-        conn_data = conn.data or {}
-        print(f"  RNS importable: {'yes' if conn_data.get('can_import_rns') else 'NO'}")
-        if conn_data.get('rns_version'):
-            print(f"  RNS version: {conn_data['rns_version']}")
-        print(f"  Config valid: {'yes' if conn_data.get('config_valid') else 'NO'}")
-        print(f"  Interfaces enabled: {conn_data.get('interfaces_enabled', 0)}")
-
-        # Collect issues and warnings from connectivity check
-        issues = list(conn_data.get('issues', []))
-        warnings = list(conn_data.get('warnings', []))
-
-        # 5. Interface dependencies
-        print("\n[5/5] Checking interface dependencies...")
-        try:
-            blocking = self._find_blocking_interfaces()
-            if blocking:
-                for iface_name, reason, fix in blocking:
-                    print(f"  ! [{iface_name}] {reason}")
-                    print(f"    Fix: {fix}")
-                    issues.append(f"Blocking interface: {iface_name}")
-            else:
-                print("  All enabled interfaces have their dependencies met")
-        except Exception as e:
-            logger.debug("Interface dependency check failed: %s", e)
-            print(f"  Could not check: {e}")
-
-        # Check if shared instance port is actually listening
-        import socket
-        try:
-            with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-                s.settimeout(1)
-                port_ok = s.connect_ex(('127.0.0.1', 37428)) == 0
-            if running and not port_ok:
-                print("  ! rnsd running but port 37428 NOT listening")
-                warnings.append("rnsd active but shared instance port not bound")
-            elif running and port_ok:
-                print(f"  Shared instance port 37428: listening")
-        except OSError:
-            pass
-
-        # Summary
-        if issues:
-            print(f"\n--- Issues Found ({len(issues)}) ---")
-            for issue in issues:
-                print(f"  ! {issue}")
-        if warnings:
-            print(f"\n--- Warnings ({len(warnings)}) ---")
-            for warning in warnings:
-                print(f"  ~ {warning}")
-
-        if not issues and not warnings:
-            print("\n--- All checks passed ---")
-        elif not issues:
-            print("\n--- Connectivity OK (with warnings) ---")
-
-        # Offer to create missing identities
-        if not identity_exists or not rns_identity.exists():
-            print("\n--- Identity Setup ---")
-            if self.dialog.yesno(
-                "Create Identities",
-                "One or more RNS identities are missing.\n\n"
-                "Create them now?\n\n"
-                "  • RNS identity: used by rnsd for network presence\n"
-                "  • Gateway identity: used by MeshForge bridge"
-            ):
-                try:
-                    result = create_identities()
-                    if result.success:
-                        print(f"  ✓ {result.message}")
-                        created = (result.data or {}).get('created', [])
-                        if 'rns' in created:
-                            print(f"    RNS identity: {result.data['rns_identity']}")
-                        if 'gateway' in created:
-                            print(f"    Gateway identity: {result.data['gateway_identity']}")
-                    else:
-                        print(f"  ✗ {result.message}")
-                except Exception as e:
-                    print(f"  ✗ Identity creation failed: {e}")
-
-        # RNS tool availability
-        print("\n--- RNS Tool Availability ---")
-        for tool in ['rnsd', 'rnstatus', 'rnpath', 'rnprobe', 'rnid', 'rncp', 'rnx']:
-            path = shutil.which(tool)
-            if path:
-                print(f"  {tool}: {path}")
-            else:
-                print(f"  {tool}: not found")
-
-        self._wait_for_enter()
-
-    def _rns_config_drift_check(self):
-        """Check for config drift between gateway and rnsd."""
-        clear_screen()
-        print("=== RNS Config Drift Check ===\n")
-        print("Comparing gateway config path vs rnsd actual path...\n")
-
-        if not _HAS_CONFIG_DRIFT:
-            print("  Config drift module not available.")
-            print("  File: src/utils/config_drift.py")
-            self._wait_for_enter()
-            return
-
-        result = detect_rnsd_config_drift()
-
-        # Display result
-        severity_colors = {
-            'info': '\033[0;34m',     # blue
-            'warning': '\033[0;33m',  # yellow
-            'error': '\033[0;31m',    # red
-        }
-        color = severity_colors.get(result.severity, '')
-        reset = '\033[0m'
-
-        if result.drifted:
-            print(f"  {color}CONFIG DRIFT DETECTED{reset}\n")
-            print(f"  Gateway resolves to: {result.gateway_config_dir}")
-            print(f"  rnsd actually uses:   {result.rnsd_config_dir}")
-            print(f"  Detection method:     {result.detection_method}")
-            if result.rnsd_pid:
-                print(f"  rnsd PID:             {result.rnsd_pid}")
-            print(f"\n  {color}Fix:{reset} {result.fix_hint}")
-        else:
-            print(f"  \033[0;32mNo drift detected\033[0m\n")
-            print(f"  {result.message}")
-            if result.gateway_config_dir:
-                print(f"  Config directory: {result.gateway_config_dir}")
-            if result.rnsd_pid:
-                print(f"  rnsd PID: {result.rnsd_pid}")
-            print(f"  Detection method: {result.detection_method}")
-
-        print()
-        self._wait_for_enter()
-
-    @staticmethod
-    def _is_root_owned_rns_config(config_path: Path) -> bool:
-        """Check if the RNS config is in a root-only location (/root/)."""
-        try:
-            return str(config_path.resolve()).startswith('/root/')
-        except OSError:
-            return str(config_path).startswith('/root/')
-
-    def _migrate_rns_config_to_etc(self, source: Path) -> bool:
-        """Migrate RNS config from root-owned location to /etc/reticulum/config.
-
-        Copies the config to /etc/reticulum/config (system-wide, preferred location),
-        sets world-readable permissions, and renames the old file to avoid confusion.
-
-        Returns True if migration succeeded.
-        """
-        target = Path('/etc/reticulum/config')
-        if target.exists():
-            self.dialog.msgbox(
-                "Cannot Migrate",
-                f"Config already exists at:\n  {target}\n\n"
-                f"Remove it first if you want to migrate from:\n  {source}"
-            )
-            return False
-        try:
-            target.parent.mkdir(parents=True, exist_ok=True)
-            shutil.copy2(str(source), str(target))
-            target.chmod(0o644)
-            # Rename old config so rnsd picks up the /etc/ one
-            backup = source.with_suffix('.migrated')
-            source.rename(backup)
-            return True
-        except (OSError, PermissionError) as e:
-            self.dialog.msgbox("Error", f"Failed to migrate config:\n{e}")
-            return False
-
-    def _deploy_rns_template(self) -> Optional[Path]:
-        """Deploy RNS template to /etc/reticulum/config (system-wide).
-
-        Returns the path where the config was deployed, or None on failure.
-        """
-        template = Path(__file__).parent.parent.parent / 'templates' / 'reticulum.conf'
-        if not template.exists():
-            return None
-
-        # Always deploy to /etc/reticulum/ (system-wide, first in search order)
-        target = Path('/etc/reticulum/config')
-        if target.exists():
-            self.dialog.msgbox(
-                "Config Exists",
-                f"Config already exists at:\n  {target}\n\n"
-                f"Use 'Edit Reticulum Config' to modify it."
-            )
-            return None
-        try:
-            target.parent.mkdir(parents=True, exist_ok=True)
-            shutil.copy2(str(template), str(target))
-            target.chmod(0o644)  # World-readable so all users and rnsd can read it
-            return target
-        except (OSError, PermissionError) as e:
-            self.dialog.msgbox("Error", f"Failed to deploy config:\n{e}")
-            return None
-
     def _repair_rns_shared_instance(self) -> bool:
         """Repair RNS shared instance — explicit user action only.
 
@@ -1098,536 +851,6 @@ class RNSMenuMixin(RNSSnifferMixin):
             # Clean up temp dir
             shutil.rmtree(tmp_dir, ignore_errors=True)
 
-    def _view_rns_config(self):
-        """View current Reticulum config."""
-        clear_screen()
-        print("=== Reticulum Configuration ===\n")
-
-        config_path = ReticulumPaths.get_config_file()
-
-        if config_path.exists():
-            # Warn if config is in root-only location
-            if self._is_root_owned_rns_config(config_path):
-                print(f"Config: {config_path}")
-                print(f"  ** This config is in /root/ - not editable without sudo **")
-                print(f"  ** Use 'Edit Reticulum Config' to migrate to /etc/reticulum/ **\n")
-            else:
-                print(f"Config: {config_path}\n")
-            try:
-                content = config_path.read_text()
-                print(content)
-
-                # Show validation warnings inline
-                issues = self._validate_rns_config_content(content)
-                if issues:
-                    print("\n--- Config Issues ---")
-                    for issue in issues:
-                        print(f"  ! {issue}")
-            except PermissionError:
-                print(f"Permission denied reading {config_path}")
-                print(f"Try: sudo cat {config_path}")
-        else:
-            print(f"No Reticulum config found at: {config_path}")
-            user_home = get_real_user_home()
-            print(f"\nMeshForge checks (in order):")
-            print(f"  1. /etc/reticulum/config  (system-wide, preferred)")
-            print(f"  2. {user_home}/.config/reticulum/config")
-            print(f"  3. {user_home}/.reticulum/config")
-            if os.geteuid() == 0 and os.environ.get('SUDO_USER'):
-                print(f"\nNote: rnsd (running as root) uses /root/.reticulum/config")
-                print(f"  For shared use, deploy to /etc/reticulum/config")
-            print(f"\nTo create: use 'Edit Reticulum Config' to deploy template")
-            print(f"Template:  templates/reticulum.conf")
-
-        # Show Meshtastic_Interface plugin status
-        plugin_path = ReticulumPaths.get_interfaces_dir() / 'Meshtastic_Interface.py'
-        print(f"\n--- Meshtastic Interface Plugin ---")
-        if plugin_path.exists():
-            print(f"  Installed: {plugin_path}")
-            print(f"  Size: {plugin_path.stat().st_size} bytes")
-        else:
-            print(f"  NOT INSTALLED")
-            print(f"  Expected at: {plugin_path}")
-            print(f"  Source: https://github.com/landandair/RNS_Over_Meshtastic")
-            print(f"  Use 'Install Meshtastic Interface' from the RNS menu to install.")
-
-        self._wait_for_enter()
-
-    def _edit_rns_config(self):
-        """Edit Reticulum config with available editor. Deploys template if no config exists."""
-        config_path = ReticulumPaths.get_config_file()
-
-        if not config_path.exists():
-            # Offer to deploy from template to /etc/reticulum/config (system-wide)
-            template = Path(__file__).parent.parent.parent / 'templates' / 'reticulum.conf'
-
-            if template.exists():
-                target = Path('/etc/reticulum/config')
-                if self.dialog.yesno(
-                    "Deploy Reticulum Config",
-                    f"No Reticulum config found.\n\n"
-                    f"Deploy template to:\n  {target}\n\n"
-                    f"This sets up RNS with:\n"
-                    f"  - share_instance = Yes (required for rnstatus)\n"
-                    f"  - AutoInterface (local network discovery)\n"
-                    f"  - Meshtastic_Interface on port 4403\n\n"
-                    f"You can edit it after deployment."
-                ):
-                    deployed = self._deploy_rns_template()
-                    if deployed:
-                        config_path = deployed
-                    else:
-                        return
-                else:  # User said No
-                    return
-            else:
-                self.dialog.msgbox(
-                    "No Config",
-                    "No Reticulum config found and template missing.\n\n"
-                    "Install RNS first: pipx install rns\n"
-                    "Then run rnsd once to generate default config."
-                )
-                return
-
-        # If config is in /root/, offer to migrate to /etc/reticulum/
-        if self._is_root_owned_rns_config(config_path):
-            if self.dialog.yesno(
-                "Migrate Config",
-                f"Config is at:\n  {config_path}\n\n"
-                f"This location requires root access to edit.\n\n"
-                f"Migrate to /etc/reticulum/config?\n"
-                f"(System-wide location, accessible by rnsd and all users)"
-            ):
-                if self._migrate_rns_config_to_etc(config_path):
-                    config_path = Path('/etc/reticulum/config')
-                    self.dialog.msgbox(
-                        "Migrated",
-                        f"Config moved to: {config_path}\n\n"
-                        f"Restart rnsd to apply:\n"
-                        f"  sudo systemctl restart rnsd"
-                    )
-                # If migration failed, continue with original path
-
-        # Find editor
-        editor = None
-        for cmd in ['nano', 'vim', 'vi']:
-            if shutil.which(cmd):
-                editor = cmd
-                break
-
-        if not editor:
-            self.dialog.msgbox("Error", "No text editor found (nano, vim, vi)")
-            return
-
-        subprocess.run([editor, str(config_path)], timeout=None)
-
-        # After editing, check for config divergence between user and root
-        self._check_rns_config_divergence(config_path)
-
-    def _check_rns_config_divergence(self, edited_path: Path):
-        """Check if edited config differs from root/system config that rnsd actually uses.
-
-        When running with sudo, user edits /home/user/.reticulum/config but
-        rnsd (running as root) reads /root/.reticulum/config. The configs
-        silently diverge, causing the user's changes to have no effect.
-
-        This check warns the user and offers to sync.
-        """
-        import os
-
-        # Only relevant when running as root/sudo
-        if os.geteuid() != 0:
-            return
-
-        # Find where rnsd actually reads its config
-        # rnsd systemd service runs as root, so it reads from one of:
-        root_configs = [
-            Path('/etc/reticulum/config'),
-            Path('/root/.config/reticulum/config'),
-            Path('/root/.reticulum/config'),
-        ]
-
-        # Skip if edited path is already a root/system path
-        edited_str = str(edited_path)
-        if edited_str.startswith('/root/') or edited_str.startswith('/etc/'):
-            return
-
-        for root_config in root_configs:
-            if root_config.exists() and root_config != edited_path:
-                # Compare contents
-                try:
-                    user_content = edited_path.read_text()
-                    root_content = root_config.read_text()
-
-                    if user_content != root_content:
-                        if self.dialog.yesno(
-                            "Config Divergence Detected",
-                            f"WARNING: Your edited config:\n"
-                            f"  {edited_path}\n\n"
-                            f"differs from the config rnsd uses:\n"
-                            f"  {root_config}\n\n"
-                            f"rnsd runs as root and reads {root_config}.\n"
-                            f"Your changes won't take effect until synced.\n\n"
-                            f"Copy your config to {root_config}?"
-                        ):
-                            try:
-                                import shutil
-                                # Backup root config first
-                                backup = root_config.with_suffix('.config.bak')
-                                if root_config.exists():
-                                    shutil.copy2(str(root_config), str(backup))
-                                shutil.copy2(str(edited_path), str(root_config))
-                                self.dialog.msgbox(
-                                    "Config Synced",
-                                    f"Copied to: {root_config}\n"
-                                    f"Backup at: {backup}\n\n"
-                                    f"Restart rnsd to apply:\n"
-                                    f"  sudo systemctl restart rnsd"
-                                )
-                            except Exception as e:
-                                self.dialog.msgbox(
-                                    "Sync Failed",
-                                    f"Could not copy config: {e}\n\n"
-                                    f"Manual fix:\n"
-                                    f"  sudo cp {edited_path} {root_config}\n"
-                                    f"  sudo systemctl restart rnsd"
-                                )
-                except (OSError, subprocess.SubprocessError) as e:
-                    logger.debug("RNS config apply failed: %s", e)
-                return  # Only check the first existing root config
-
-    def _validate_rns_config_content(self, content: str) -> list:
-        """Validate RNS config content and return list of issues found.
-
-        Checks for common misconfigurations that cause rnstatus/rnpath failures:
-        - Missing [reticulum] section
-        - Missing share_instance = Yes (required for client apps to connect)
-        - No interfaces configured
-        - No Meshtastic_Interface (needed for mesh bridging)
-        - Meshtastic_Interface.py plugin not installed
-        """
-        issues = []
-        content_lower = content.lower()
-
-        # Check [reticulum] section exists
-        if '[reticulum]' not in content_lower:
-            issues.append("Missing [reticulum] section")
-
-        # Check share_instance (required for rnstatus/rnpath to connect to rnsd)
-        has_share = False
-        for line in content.split('\n'):
-            stripped = line.strip().lower()
-            if stripped.startswith('#'):
-                continue
-            if 'share_instance' in stripped:
-                if 'yes' in stripped or 'true' in stripped:
-                    has_share = True
-                break
-        if not has_share:
-            issues.append("share_instance not set to Yes (rnstatus/client apps won't connect)")
-
-        # Check for at least one active interface
-        has_interface = False
-        has_meshtastic = False
-        for line in content.split('\n'):
-            stripped = line.strip()
-            if stripped.startswith('#'):
-                continue
-            if stripped.startswith('[[') and stripped.endswith(']]'):
-                has_interface = True
-            if 'meshtastic_interface' in stripped.lower() and 'type' in stripped.lower():
-                has_meshtastic = True
-
-        if not has_interface:
-            issues.append("No interfaces configured")
-
-        # Check Meshtastic_Interface status: config reference + plugin file
-        plugin_path = ReticulumPaths.get_interfaces_dir() / 'Meshtastic_Interface.py'
-        plugin_installed = plugin_path.exists()
-
-        if not has_meshtastic and not plugin_installed:
-            issues.append("No Meshtastic_Interface configured (needed for mesh bridging)")
-        elif has_meshtastic and not plugin_installed:
-            issues.append(
-                f"Meshtastic_Interface.py plugin not installed at "
-                f"{ReticulumPaths.get_interfaces_dir()}/\n"
-                f"    Install from: https://github.com/landandair/RNS_Over_Meshtastic"
-            )
-
-        return issues
-
-    def _check_rns_setup(self) -> bool:
-        """Check RNS setup and offer to fix common issues.
-
-        Available via 'Check RNS Setup' menu item. Returns True if setup
-        looks OK or user chose to continue anyway, False if user wants
-        to go back.
-        """
-        config_path = ReticulumPaths.get_config_file()
-
-        if not config_path.exists():
-            template = Path(__file__).parent.parent.parent / 'templates' / 'reticulum.conf'
-            if template.exists():
-                target = Path('/etc/reticulum/config')
-                if self.dialog.yesno(
-                    "RNS Not Configured",
-                    f"No Reticulum config found.\n\n"
-                    f"RNS tools (rnstatus, rnpath) and the gateway bridge\n"
-                    f"require a config file to function.\n\n"
-                    f"Deploy MeshForge template to:\n"
-                    f"  {target}\n\n"
-                    f"(Sets up shared instance + Meshtastic bridge)"
-                ):
-                    deployed = self._deploy_rns_template()
-                    if deployed:
-                        self.dialog.msgbox(
-                            "Config Deployed",
-                            f"Deployed to: {deployed}\n\n"
-                            f"Restart rnsd to apply:\n"
-                            f"  sudo systemctl restart rnsd"
-                        )
-                        config_path = deployed
-            return True  # Continue to menu either way
-
-        # Config exists - check if it's in a root-only location
-        if self._is_root_owned_rns_config(config_path):
-            if self.dialog.yesno(
-                "Config in /root/",
-                f"RNS config found at:\n  {config_path}\n\n"
-                f"This location requires root access to edit.\n\n"
-                f"Migrate to /etc/reticulum/config?\n"
-                f"(System-wide location, accessible by all users)"
-            ):
-                if self._migrate_rns_config_to_etc(config_path):
-                    config_path = Path('/etc/reticulum/config')
-                    self.dialog.msgbox(
-                        "Migrated",
-                        f"Config moved to: {config_path}\n\n"
-                        f"Restart rnsd to apply:\n"
-                        f"  sudo systemctl restart rnsd"
-                    )
-
-        # Config exists - validate it
-        try:
-            content = config_path.read_text()
-            issues = self._validate_rns_config_content(content)
-            if issues:
-                msg = f"Config: {config_path}\n\nIssues found:\n"
-                for issue in issues:
-                    msg += f"  - {issue}\n"
-                msg += f"\nUse 'Edit Reticulum Config' to fix these issues."
-                self.dialog.msgbox("RNS Config Issues", msg)
-        except PermissionError:
-            self.dialog.msgbox(
-                "Permission Denied",
-                f"Cannot read config at:\n  {config_path}\n\n"
-                f"Run MeshForge with sudo to access this file,\n"
-                f"or use 'Edit Reticulum Config' to migrate it."
-            )
-
-        # Check for Meshtastic_Interface.py plugin (separate from config validation)
-        if not self._check_meshtastic_plugin():
-            if self.dialog.yesno(
-                "Meshtastic Interface Plugin Missing",
-                "The Meshtastic_Interface.py plugin is not installed.\n\n"
-                "This plugin is required for bridging RNS over\n"
-                "Meshtastic LoRa mesh networks.\n\n"
-                f"Expected at:\n"
-                f"  {ReticulumPaths.get_interfaces_dir()}/Meshtastic_Interface.py\n\n"
-                "Download and install it now?"
-            ):
-                self._install_meshtastic_interface_plugin()
-
-        return True
-
-    def _run_rns_tool(self, cmd: list, tool_name: str):
-        """Run an RNS CLI tool with address-in-use error detection.
-
-        Captures both stdout and stderr to detect specific error patterns.
-        RNS logs errors to stdout in some configurations, so both streams
-        must be checked for the 'Address already in use' pattern.
-
-        Args:
-            cmd: Command and arguments to run
-            tool_name: Display name for error messages (e.g., "rnpath")
-        """
-        try:
-            result = subprocess.run(
-                cmd, capture_output=True, text=True, timeout=15
-            )
-            # RNS tools may log errors to stdout or stderr depending on config
-            combined = (result.stdout or "") + (result.stderr or "")
-
-            if result.returncode == 0:
-                # Success - show normal output
-                if result.stdout:
-                    print(result.stdout, end='')
-            elif "address already in use" in combined.lower():
-                # Suppress noisy traceback, show actionable diagnostics
-                print("\nError: RNS port conflict (Address already in use)")
-                print("Another process is bound to the RNS AutoInterface port.\n")
-                self._diagnose_rns_port_conflict()
-            elif "no shared" in combined.lower() or "could not connect" in combined.lower() or "could not get" in combined.lower() or "shared instance" in combined.lower() or "authenticationerror" in combined.lower() or "digest" in combined.lower():
-                # RNS shared instance issue — DIAGNOSE, don't auto-fix.
-                # Auto-fix was the #1 source of regressions (see persistent_issues.md).
-                # Policy: show what's wrong, let the user decide what to do.
-                print(f"\nRNS connectivity issue detected.")
-
-                # Check if rnsd is actually running
-                rnsd_running = False
-                try:
-                    r = subprocess.run(
-                        ['systemctl', 'is-active', 'rnsd'],
-                        capture_output=True, text=True, timeout=5
-                    )
-                    rnsd_running = r.stdout.strip() == 'active'
-                except (subprocess.SubprocessError, OSError):
-                    pass
-
-                if not rnsd_running:
-                    # rnsd is NOT running — offer to start it (with user consent)
-                    print("rnsd is not running.\n")
-                    if self.dialog.yesno(
-                        "rnsd Not Running",
-                        f"{tool_name} failed because rnsd is not running.\n\n"
-                        "Start rnsd now?\n\n"
-                        "If rnsd won't start, use RNS > Diagnostics to investigate.",
-                    ):
-                        try:
-                            subprocess.run(
-                                ['systemctl', 'start', 'rnsd'],
-                                capture_output=True, text=True, timeout=30
-                            )
-                            print("Starting rnsd...")
-                            if self._wait_for_rns_port(max_wait=10):
-                                print(f"rnsd started. Retrying {tool_name}...\n")
-                                retry_result = subprocess.run(
-                                    cmd, capture_output=True, text=True, timeout=15
-                                )
-                                if retry_result.returncode == 0 and retry_result.stdout:
-                                    print(retry_result.stdout, end='')
-                                else:
-                                    self._diagnose_rns_connectivity(
-                                        (retry_result.stdout or "") + (retry_result.stderr or "")
-                                    )
-                            else:
-                                print("rnsd started but port 37428 not listening.")
-                                print("Check: sudo journalctl -u rnsd -n 20")
-                                print("Or run: RNS > Diagnostics from the menu.")
-                        except (subprocess.SubprocessError, OSError) as e:
-                            print(f"Failed to start rnsd: {e}")
-                    else:
-                        print("To start rnsd manually: sudo systemctl start rnsd")
-                else:
-                    # rnsd IS running but tools can't connect.
-                    # Most common cause: rnsd still initializing (crypto, interfaces).
-                    # Wait for port 37428 before showing diagnostics.
-                    print("rnsd is running — waiting for port 37428...")
-                    port_ready = self._wait_for_rns_port()
-                    if port_ready:
-                        # Port came up — retry the tool
-                        print(f"Port ready. Retrying {tool_name}...\n")
-                        retry_result = subprocess.run(
-                            cmd, capture_output=True, text=True, timeout=15
-                        )
-                        if retry_result.returncode == 0 and retry_result.stdout:
-                            print(retry_result.stdout, end='')
-                        else:
-                            # Port is up but tool still fails — auth or config issue
-                            self._diagnose_rns_connectivity(
-                                (retry_result.stdout or "") + (retry_result.stderr or "")
-                            )
-                    else:
-                        # Port never came up — show diagnostics
-                        self._diagnose_rns_connectivity(combined)
-            else:
-                # Other error - DON'T auto-fix, just show output
-                # RNS tools may return non-zero for benign reasons (empty table, no paths)
-                if result.stdout:
-                    print(result.stdout, end='')
-                if result.stderr and result.stderr.strip():
-                    # Only show stderr if it contains actual error info
-                    stderr_lower = result.stderr.lower()
-                    if "error" in stderr_lower or "failed" in stderr_lower or "exception" in stderr_lower:
-                        print(f"\nNote: {tool_name} reported an issue:")
-                        for line in result.stderr.strip().split('\n')[-3:]:
-                            print(f"  {line}")
-        except FileNotFoundError:
-            print(f"\n{tool_name} not found. Is RNS installed?")
-            print("Install: pipx install rns")
-        except subprocess.TimeoutExpired:
-            print(f"\n{tool_name} timed out. RNS may be unresponsive.")
-            print("Try restarting rnsd: sudo systemctl restart rnsd")
-
-    def _wait_for_rns_port(self, max_wait: int = 10) -> bool:
-        """Wait for rnsd to start listening on port 37428.
-
-        Polls the port with 1-second intervals. Returns True if port
-        becomes available, False if timeout expires.
-        """
-        import socket
-        for i in range(max_wait):
-            try:
-                with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-                    s.settimeout(1)
-                    if s.connect_ex(('127.0.0.1', 37428)) == 0:
-                        return True
-            except OSError:
-                pass
-            time.sleep(1)
-        return False
-
-    def _diagnose_rns_connectivity(self, error_output: str):
-        """Show targeted diagnostics when rnsd is running but tools can't connect.
-
-        Instead of guessing, check for auth errors (actionable) then fall
-        through to showing the actual rnsd journal log.
-        """
-        lower = error_output.lower()
-        print("rnsd is running but RNS tools cannot connect.\n")
-
-        # Auth errors are actionable — detect and show specific fix
-        if "authenticationerror" in lower or "digest" in lower:
-            print("Cause: RPC authentication mismatch (stale auth tokens)")
-            print("Fix:   Clear auth tokens and restart rnsd:\n")
-            print("  sudo systemctl stop rnsd")
-            print("  sudo rm -f /etc/reticulum/storage/shared_instance_*")
-            print("  sudo rm -f /root/.reticulum/storage/shared_instance_*")
-            user_home = get_real_user_home()
-            print(f"  rm -f {user_home}/.reticulum/storage/shared_instance_*")
-            print("  sudo systemctl start rnsd")
-            return
-
-        # Check for blocking interfaces (most common root cause)
-        blocking = self._find_blocking_interfaces()
-        if blocking:
-            print("Cause: rnsd is stuck initializing a blocking interface.\n")
-            for iface_name, reason, fix in blocking:
-                print(f"  [{iface_name}] {reason}")
-                print(f"  Fix: {fix}\n")
-            print("Options:")
-            print("  1. Start the missing dependency (see Fix above)")
-            print("  2. Disable the interface in /etc/reticulum/config")
-            print("     (change 'enabled = yes' to 'enabled = no')")
-            print("  3. sudo systemctl restart rnsd (after fixing)")
-            return
-
-        # No specific cause detected — show actual rnsd log
-        print("Showing recent rnsd log:\n")
-        try:
-            r = subprocess.run(
-                ['journalctl', '-u', 'rnsd', '-n', '15', '--no-pager'],
-                capture_output=True, text=True, timeout=10
-            )
-            if r.stdout and r.stdout.strip():
-                for line in r.stdout.strip().split('\n'):
-                    print(f"  {line}")
-            else:
-                print("  (no log output)")
-        except (subprocess.SubprocessError, OSError):
-            print("  (could not read journal)")
-        print("\nTo restart: sudo systemctl restart rnsd")
-
     def _find_blocking_interfaces(self) -> list:
         """Check if enabled RNS interfaces have missing dependencies.
 
@@ -1652,7 +875,6 @@ class RNSMenuMixin(RNSSnifferMixin):
 
         # Parse enabled interfaces from the config
         # RNS config uses [[InterfaceName]] sections with type= and enabled=
-        import re
         # Match interface sections: [[Name]] ... type = ... enabled = yes
         iface_pattern = re.compile(
             r'^\s*\[\[(.+?)\]\]\s*$'
@@ -1809,96 +1031,11 @@ class RNSMenuMixin(RNSSnifferMixin):
 
         return disabled
 
-    def _check_nomadnet_conflict(self) -> bool:
-        """Check if NomadNet is running and holding the shared instance port.
-
-        NomadNet creates its own Reticulum() instance and becomes the shared
-        instance on port 37428. If rnsd is also configured with
-        share_instance = Yes, they fight over the port causing crash loops.
-
-        Returns True if NomadNet conflict detected.
-        """
-        try:
-            result = subprocess.run(
-                ['pgrep', '-f', 'nomadnet'],
-                capture_output=True, text=True, timeout=5
-            )
-            return result.returncode == 0
-        except (subprocess.SubprocessError, OSError):
-            return False
-
-    def _diagnose_rns_port_conflict(self):
-        """Diagnose and offer to fix RNS port conflicts from the TUI."""
-        import time
-        try:
-            # Check NomadNet first — most common cause of port conflicts
-            if self._check_nomadnet_conflict():
-                print("CAUSE: NomadNet is running and owns port 37428.")
-                print("rnsd can't start because NomadNet has the port.\n")
-
-                if self.dialog.yesno(
-                    "Fix Port Conflict",
-                    "NomadNet is holding port 37428.\n\n"
-                    "MeshForge can fix this:\n"
-                    "  1. Stop NomadNet\n"
-                    "  2. Start rnsd (becomes shared instance)\n"
-                    "  3. Restart NomadNet (connects as client)\n\n"
-                    "Fix now?"
-                ):
-                    print("Stopping NomadNet...")
-                    subprocess.run(
-                        ['pkill', '-f', 'nomadnet'],
-                        capture_output=True, timeout=5
-                    )
-                    time.sleep(1)
-
-                    print("Starting rnsd...")
-                    subprocess.run(
-                        ['systemctl', 'start', 'rnsd'],
-                        capture_output=True, text=True, timeout=15
-                    )
-                    time.sleep(2)
-
-                    print("Restarting NomadNet as client...")
-                    subprocess.run(
-                        ['systemctl', '--user', 'start', 'nomadnet'],
-                        capture_output=True, text=True, timeout=10
-                    )
-                    print("Done. Startup order: rnsd -> NomadNet -> MeshForge\n")
-                return
-
-            # Use centralized service check when available
-            if _HAS_SERVICE_CHECK:
-                rnsd_running = check_process_running('rnsd')
-            else:
-                # Fallback to direct pgrep
-                result = subprocess.run(
-                    ['pgrep', '-f', 'rnsd'],
-                    capture_output=True, text=True, timeout=5
-                )
-                rnsd_running = result.returncode == 0
-
-            if rnsd_running:
-                # Get PID for diagnostic message
-                try:
-                    pid_result = subprocess.run(
-                        ['pgrep', '-f', 'rnsd'],
-                        capture_output=True, text=True, timeout=5
-                    )
-                    pid = pid_result.stdout.strip().split('\n')[0] if pid_result.stdout else 'unknown'
-                except (subprocess.SubprocessError, OSError) as e:
-                    logger.debug("rnsd PID lookup failed: %s", e)
-                    pid = 'unknown'
-                print(f"rnsd is running (PID: {pid}) but may need a restart:")
-                print("  sudo systemctl restart rnsd")
-            else:
-                print("No rnsd found. A stale process may be holding the port.")
-                print("  Find it:    sudo lsof -i UDP:29716")
-                print("  Kill stale: pkill -f rnsd")
-                print("  Or wait ~30s for the socket to timeout")
-        except (subprocess.SubprocessError, OSError) as e:
-            logger.debug("RNS port conflict diagnosis failed: %s", e)
-            print("  Try: sudo systemctl restart rnsd")
-
     # Sniffer methods (_rns_traffic_sniffer, _rns_sniffer_*) are inherited
     # from RNSSnifferMixin - see rns_sniffer_mixin.py
+    #
+    # Config methods (_view_rns_config, _edit_rns_config, etc.) are inherited
+    # from RNSConfigMixin - see rns_config_mixin.py
+    #
+    # Diagnostics methods (_rns_diagnostics, _run_rns_tool, etc.) are inherited
+    # from RNSDiagnosticsMixin - see rns_diagnostics_mixin.py
